moved check_status_api_events api to dnac.py common helper file
